pyproject.toml
[project]
name = "feyipay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-filter>=25.2",
    "djangorestframework-simplejwt>=5.5.1",
    "markdown>=3.10.2",
    "mysqlclient>=2.2.8",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
